aimp_testbed/requirements.txt
msgpack==1.0.8
pynacl==1.5.0
